# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Flask Terraform API",
    "description": "API to manage Terraform infrastructure",
    "version": "1.0.0"
  },
  "paths": {
    "/terraform/ec2": {
      "post": {
        "summary": "Launch an EC2 instance using Terraform",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "EC2 instance launched",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "public_ip": { "type": "string" },
                    "vpc_id": { "type": "string" },
                    "subnet_id": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/terraform/openstack": {
      "post": {
        "summary": "Launch an OpenStack instance using Terraform",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OpenStack instance launched",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "public_ip": { "type": "string" },
                    "vpc_id": { "type": "string" },
                    "subnet_id": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/terraform/status": {
      "get": {
        "summary": "Get instance status",
        "responses": {
          "200": {
            "description": "Instance details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "public_ip": { "type": "string" },
                    "vpc_id": { "type": "string" },
                    "subnet_id": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/launch": {
      "get": {
        "summary": "Launch instance in browser",
        "responses": {
          "302": {
            "description": "Redirect to the launched instance"
          }
        }
      }
    },
    "/deploy/fedora": {
      "post": {
        "summary": "Deploy Fedora Server with Ansible",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "server_ip": { "type": "string" },
                  "ssh_user": { "type": "string" },
                  "ssh_password": { "type": "string" },
                  "libraries": {
                    "type": "array",
                    "items": { "type": "string" }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Fedora server deployed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "redirect_url": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/redirect/jupyter": {
      "get": {
        "summary": "Redirect to Jupyter Notebook on the deployed Fedora Server",
        "parameters": [
          {
            "name": "server_ip",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to Jupyter Notebook"
          }
        }
      }
    },
    "/test/local": {
  "post": {
    "summary": "Run a Jupyter Notebook locally",
    "requestBody": {
      "required": true,
      "content": {
        "multipart/form-data": {
          "schema": {
            "type": "object",
            "properties": {
              "notebook_file": {
                "type": "string",
                "format": "binary",
                "description": "The Jupyter Notebook (.ipynb) file to be executed locally."
              },
              "model_name": {
                "type": "string",
                "description": "The name of the model to be tested within the notebook."
              }
            }
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "Notebook executed successfully",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" },
                "executed_notebook": { "type": "string", "description": "Download URL for the executed notebook." }
              }
            }
          }
        }
      },
      "400": {
        "description": "Invalid request data"
      },
      "500": {
        "description": "Error executing the notebook"
      }
    }
  }
}

  }
}